# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SiteIQ Construction Site Coordination API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workers:
    get:
      operationId: listWorkers
      tags: [workers]
      summary: List all workers
      parameters:
        - name: zone
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Workers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Worker"
    post:
      operationId: createWorker
      tags: [workers]
      summary: Create a worker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkerInput"
      responses:
        "201":
          description: Worker created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Worker"
  /workers/{id}:
    get:
      operationId: getWorker
      tags: [workers]
      summary: Get worker by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Worker
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Worker"
    patch:
      operationId: updateWorker
      tags: [workers]
      summary: Update worker
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkerUpdate"
      responses:
        "200":
          description: Worker updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Worker"
    delete:
      operationId: deleteWorker
      tags: [workers]
      summary: Delete worker
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Worker:
      type: object
      required: [id, name, trade, status, zone, ppeScore, fatigueScore, tenantId]
      properties:
        id:
          type: integer
        name:
          type: string
        trade:
          type: string
        certifications:
          type: array
          items:
            type: string
        zone:
          type: string
        status:
          type: string
        ppeScore:
          type: integer
        fatigueScore:
          type: integer
        locationX:
          type: ["number", "null"]
        locationY:
          type: ["number", "null"]
        shiftStart:
          type: ["string", "null"]
        userId:
          type: ["integer", "null"]
        tenantId:
          type: integer
    WorkerInput:
      type: object
      required: [name, trade, zone]
      properties:
        name:
          type: string
        trade:
          type: string
        certifications:
          type: array
          items:
            type: string
        zone:
          type: string
        status:
          type: string
        ppeScore:
          type: integer
        fatigueScore:
          type: integer
        locationX:
          type: number
        locationY:
          type: number
        shiftStart:
          type: string
    WorkerUpdate:
      type: object
      properties:
        name:
          type: string
        trade:
          type: string
        zone:
          type: string
        status:
          type: string
        ppeScore:
          type: integer
        fatigueScore:
          type: integer
        locationX:
          type: number
        locationY:
          type: number
...